+++ docs/dbt_runner/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this gear are documented in this file.
 ## 0.2.0
 
 * Refactors to support multiple source prefixes and pull directly from S3 instead of relying on FW Storage/Datasets
+* Only save DBT artifacts if running in `debug` mode
 
 ## 0.1.1